requirements.txt
llama-cpp-python
llama-cpp-python[server]
discord.py
httpx
pillow
requests
aiohttp
xtts-api-server
pytesseract